refactor: moved modified auto-gen code
