+++ pslab-core.X/main.c
@@ -1,15 +1,8 @@
 #include "mcc_generated_files/system.h"
 #include "mcc_generated_files/fatfs/fatfs_demo.h"
-#include "mcc_generated_files/uart1.h"
+#include "protocols/uart1.h"
 
 #include "states.h"
-
-/**
- * @brief: Discard incoming serial traffic.
- */
-void ClearBuffer(void) {
-    while (UART1_IsRxReady()) UART1_Read();
-}
 
 int main(void) {
     SYSTEM_Initialize();
